Qt: Make Log and Log Configuration widget separate
