+++ Source/Core/DolphinQt2/CMakeLists.txt
@@ -43,7 +43,8 @@ set(SRCS
   Config/Graphics/GraphicsWindow.cpp
   Config/Graphics/SoftwareRendererWidget.cpp
   Config/InfoWidget.cpp
-  Config/LoggerWidget.cpp
+  Config/LogConfigWidget.cpp
+  Config/LogWidget.cpp
   Config/Mapping/GCKeyboardEmu.cpp
   Config/Mapping/GCPadEmu.cpp
   Config/Mapping/GCPadWiiU.cpp
